Remove Options command & add Preferences command
